+++ src/components/answer.tsx
@@ -7,7 +7,7 @@ import {
 import { ScenarioData } from "@/models/scenario";
 import { ChevronsUpDown, CircleCheckBig, CircleX } from "lucide-react";
 import { useState } from "react";
-import { Conversation } from "@/components/conversation/conversation";
+import { Conversation } from "@/conversation";
 import { Craft } from "@/components/craft/craft";
 import { Problems } from "@/components/problems";
 import { Button } from "@/components/ui/button";
