--- src/components/conversation/conversation.css
@@ -1,33 +1,0 @@
-@import "tailwindcss";
-
-.chat-row {
-  @apply flex mt-3 mb-3;
-}
-
-.chat-row.left {
-  @apply justify-start;
-}
-
-.chat-row.right {
-  @apply justify-end;
-}
-
-.chat-bubble {
-  @apply max-w-[70%];
-}
-
-.label {
-  @apply text-sm font-bold mb-1;
-}
-
-.bubble {
-  @apply px-4 py-3 rounded-[1em] leading-[1.4] break-words whitespace-pre-line;
-}
-
-.pilot .bubble {
-  @apply bg-[var(--color-chat-pilot)] text-[var(--color-chat-pilot-foreground)] rounded-tl-none;
-}
-
-.controller .bubble {
-  @apply bg-[var(--color-chat-controller)] text-[var(--color-chat-controller-foreground)] rounded-tl-none;
-}
